requirements.txt
torchvision
numpy
configparser